northwind-mcp-client/requirements.txt
python-dotenv
streamlit
requests
langchain
langchain-openai
openai
pytest
